// src/store/notifications/notifications.enum.ts
enum NotificationsActionTypeEnum {

    GET_NOTIFICATIONS_PENDING = "GET_NOTIFICATIONS_PENDING",
    GET_NOTIFICATIONS_SUCCESS = "GET_NOTIFICATIONS_SUCCESS",
    GET_NOTIFICATIONS_FAILED = "GET_NOTIFICATIONS_FAILED",

    DELETE_NOTIFICATION_SUCCESS = "DELETE_NOTIFICATION_SUCCESS",
    DELETE_NOTIFICATION_FAILED = "DELETE_NOTIFICATION_FAILED",
    DELETE_NOTIFICATION_PENDING = "DELETE_NOTIFICATION_PENDING",

    UPDATE_NOTIFICATION_SUCCESS = "UPDATE_NOTIFICATION_SUCCESS",
    UPDATE_NOTIFICATION_FAILED = "UPDATE_NOTIFICATION_FAILED",
    UPDATE_NOTIFICATION_PENDING = "UPDATE_NOTIFICATION_PENDING",
}

export default NotificationsActionTypeEnum;